Before Azure connection tested

+++ src/main/java/bugBust/transitgo/repository/BusRouteRepository.java
@@ -1,8 +1,7 @@
 package bugBust.transitgo.repository;
-
 import bugBust.transitgo.model.BusRoute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusRouteRepository extends JpaRepository<BusRoute,Long> {
-BusRoute getByrouteno(Integer routeno);
+public interface BusRouteRepository extends JpaRepository<BusRoute,Integer> {
+BusRoute getByrouteno(int routeno);
 }
